requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
